Fix whole-library review findings across all backends
Exhaustive fresh-eyes review of the entire library (correctness, security,
memory/concurrency, performance) surfaced 127 confirmed findings; all are
fixed here with full corrections (no workarounds, no CPU fallbacks) and
verified on CPU, CUDA (RTX 5070), ROCm (gfx1150), OneAPI, and Vulkan.

Highlights by subsystem:
- autograd: use-after-free in shape_ext index paths, bounded slice-index LRU,
  true forward-mode jacfwd via jvp.
- nn: LPPool averaging corrected (sum of |x|^p) across all 5 backends;
  conv_transpose output_padding/dilation attrs; attention position_bias
  train/eval parity.
- CPU kernels: LayerNorm/RMSNorm Float64 accumulators; AVX2 int8 GEMM widen
  path; AVX512 integer bit-ops.
- OneAPI: bitonic sort range, layernorm-backward double-weight, scatter
  shape contract, batched linalg.
- CUDA: flash-attention __syncthreads, cooperative nested LayerNorm, scatter
  contract.
- ROCm: grouped Conv3d in-channel indexing, LSTM device-math, flash-attn LDS
  sizing, IEEE-754 NaN patterns.
- Vulkan: caching-allocator block-merge buffer recreation; per-dtype pooling
  shaders; new cdist/masked_scatter Float64 shaders.
- jit/mlir: causal-mask KV-cache offset, deterministic constant emission.
- ONNX exporter: dedup weights so JIT-traced and structural exports no longer
  serialize each parameter twice (halves .onnx size); regression test added.
- core/serving/python: promote_types F16+BF16, memory-manager, bindings
  bounds/step handling, allocator overflow guards.

Also corrected two tests that encoded the bugs they covered (LPPool
averaging; OneAPI scatter shape gap). Affected test suites pass.

+++ src/backends/cuda/kernels/gru.cu
@@ -59,7 +59,7 @@ __global__ void gru_cell_forward_fused(
     int64_t batch_size,
     int64_t hidden_size) {
 
-    int64_t idx = blockIdx.x * blockDim.x + threadIdx.x;
+    int64_t idx = static_cast<int64_t>(blockIdx.x) * blockDim.x + threadIdx.x;
     int64_t total_elements = batch_size * hidden_size;
 
     if (idx < total_elements) {
@@ -119,7 +119,7 @@ __global__ void gru_cell_backward_fused(
     int64_t batch_size,
     int64_t hidden_size) {
 
-    int64_t idx = blockIdx.x * blockDim.x + threadIdx.x;
+    int64_t idx = static_cast<int64_t>(blockIdx.x) * blockDim.x + threadIdx.x;
     int64_t total_elements = batch_size * hidden_size;
 
     if (idx < total_elements) {
